Fix broken functional test
After merging 0.3.9 into develop, we found that the
test_submit_and_retrieve_message functional test was broken. The
breakage was intermittent but frequent, and manifested as failure to
find target elements by class name or CSS selector in a target page.
When we inspected the source of the target page during a failing test,
we found that it instead contained the XML source for a Firefox
"Connection Interrupted" error page instead of the expected HTML from
the SecureDrop Source Interface web application. We also noticed that
this tended to occur in conjunction with a "Broken Pipe" traceback from
the Source Interface server started by the functional test runner.

This breakage was unexpected because the full test suite is passing on
both the develop branch and on the 0.3.9 tag.

Our current theory is that this error was caused by Selenium making
multiple concurrent requests to the Werkzeug toy web server, which does
not handle concurrent requests well. For our purposes, it seems
sufficient to tell Werkzeug to run the test web server with
`threaded=True` so it can better handle concurrent requests, as
suggested by this Stack Overflow poster:
http://stackoverflow.com/a/12591943.

We are not sure why merging 0.3.9 would have caused this behavior to
change, but theorize that it could be due to recent changes in Firefox,
which has been updated recently, or due to recent changes in Selenium or
Flask, which were both updated as part of 0.3.9.